backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi",
    "httpx>=0.28.1",
    "langchain",
    "langchain-chroma",
    "langchain-classic>=1.0.1",
    "langchain-community",
    "langchain-core",
    "langchain-google-genai",
    "langchain-huggingface",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.5",
    "pydantic",
    "python-dotenv",
    "python-multipart>=0.0.21",
    "sentence-transformers",
    "supabase",
    "tavily-python>=0.7.19",
    "uvicorn",
]
